Fixing the Home and Expenses pages
